+++ ek-computation/src/ffn/expert_torch.rs
@@ -1,152 +1,18 @@
-use std::{
-    borrow::Borrow,
-    sync::{Arc, Mutex},
-};
+use std::sync::{Arc, Mutex};
 
-use crate::{
-    tch_safetensors::{dtype_to_tch_kind, tch_kind_to_dtype, write_safetensors},
-    x,
-};
+use crate::backend::{DType, Device, EkTensor, torch::TchTensor};
 
 use ek_base::error::EKResult;
 use once_cell::sync::OnceCell;
-use safetensors::tensor::TensorView;
 use tch::{
-    self, Tensor,
+    self,
     nn::{self, Module},
 };
 
-use super::{DType, Device, EkTensor, Expert, ExpertShape, ExpertWeight, FromSafeTensor};
-
-pub struct TchTensor(Tensor);
-
-impl From<tch::Tensor> for TchTensor {
-    fn from(value: tch::Tensor) -> Self {
-        TchTensor(value)
-    }
-}
-
-impl Borrow<Tensor> for TchTensor {
-    fn borrow(&self) -> &Tensor {
-        &self.0
-    }
-}
-
-impl From<DType> for tch::Kind {
-    fn from(val: DType) -> Self {
-        match val {
-            DType::Uint8 => tch::Kind::Uint8,
-            DType::Int16 => tch::Kind::Int16,
-            DType::Int8 => tch::Kind::Int8,
-            DType::BFloat16 => tch::Kind::BFloat16,
-            DType::Float8e4m3fn => tch::Kind::Float8e4m3fn,
-            DType::Float8e4m3fnuz => tch::Kind::Float8e4m3fnuz,
-            DType::Float => tch::Kind::Float,
-        }
-    }
-}
-
-impl From<Device> for tch::Device {
-    fn from(val: Device) -> Self {
-        match val {
-            Device::CPU => tch::Device::Cpu,
-        }
-    }
-}
-
-struct TchSafeView<'a> {
-    tensor: &'a Tensor,
-    shape: Vec<usize>,
-    dtype: safetensors::Dtype,
-}
-
-impl safetensors::View for TchSafeView<'_> {
-    fn dtype(&self) -> safetensors::Dtype {
-        self.dtype
-    }
-    fn shape(&self) -> &[usize] {
-        &self.shape
-    }
-
-    fn data(&self) -> std::borrow::Cow<[u8]> {
-        let mut data = vec![0; self.data_len()];
-        let numel = self.tensor.numel();
-        self.tensor.f_copy_data_u8(&mut data, numel).unwrap();
-        data.into()
-    }
-
-    fn data_len(&self) -> usize {
-        self.tensor.numel() * self.tensor.kind().elt_size_in_bytes()
-    }
-}
-
-impl<'a> From<&'a TchTensor> for TchSafeView<'a> {
-    fn from(val: &'a TchTensor) -> Self {
-        let dtype = tch_kind_to_dtype(val.0.kind()).unwrap();
-        let shape = val.0.size().iter().map(|&x| x as usize).collect();
-        Self {
-            tensor: &val.0,
-            shape,
-            dtype,
-        }
-    }
-}
-
-impl EkTensor for TchTensor {
-    fn rand(shape: Vec<usize>, typ: DType, dev: Device) -> Self {
-        let rand = Tensor::randn(
-            shape.into_iter().map(|x| x as i64).collect::<Vec<i64>>(),
-            (typ.into(), dev.into()),
-        );
-        TchTensor(rand)
-    }
-
-    fn stack(tensors: &[Self], dim: usize) -> Self {
-        TchTensor(tch::Tensor::stack(tensors, dim as i64))
-    }
-
-    fn shape(&self) -> Vec<usize> {
-        self.0.size().iter().map(|&x| x as usize).collect()
-    }
-
-    fn serialize(&self) -> Vec<u8> {
-        write_safetensors(&[("data", &self.0)]).unwrap()
-    }
-
-    fn from_raw(data: &[u8], shape: &[usize], dtype: DType) -> Self {
-        unsafe {
-            Tensor::from_blob(
-                data.as_ptr(),
-                &shape.iter().map(|x| *x as i64).collect::<Vec<i64>>(),
-                &[],
-                dtype.into(),
-                tch::Device::Cpu,
-            )
-            .into()
-        }
-    }
-
-    fn from_tensor_view(tv: &TensorView<'_>) -> Self {
-        tv.into()
-    }
-}
-
-impl FromSafeTensor for TchTensor {}
-
-impl From<&TensorView<'_>> for TchTensor {
-    fn from(value: &TensorView<'_>) -> Self {
-        let size: Vec<i64> = value.shape().iter().map(|&x| x as i64).collect();
-        let kind: tch::Kind = dtype_to_tch_kind(value.dtype()).unwrap();
-        let t = Tensor::f_from_data_size(value.data(), &size, kind).unwrap();
-        TchTensor(t)
-    }
-}
-
-impl TchTensor {
-    pub fn inner(&self) -> Tensor {
-        self.0.shallow_clone()
-    }
-}
+use super::{
+    ExpertWeight,
+    meta::{Expert, ExpertShape},
+};
 
 pub struct TorchFFN {
     dim: usize,
@@ -189,31 +55,25 @@ pub fn w8a16_activate(x: &tch::Tensor, s: &tch::Tensor, block_size: i64) -> tch:
 unsafe impl Sync for TorchFFN {}
 
 impl TorchFFN {
-    pub fn new(inst: x::EKInstance) -> Self {
-        let weight =
-            ExpertWeight::from_rand(inst.hidden, inst.intermediate, DType::BFloat16, Device::CPU);
-        Self::construct(inst, weight).unwrap()
-    }
-
     pub fn load_module(&self) -> Arc<Mutex<nn::Sequential>> {
         let m = self.module.get_or_init(|| {
             tch::no_grad(|| {
                 let w1_tensor = self
                     .weight
                     .up_w
-                    .0
+                    .inner()
                     .shallow_clone()
                     .to_kind(tch::Kind::BFloat16);
                 let w2_tensor = self
                     .weight
                     .down_w
-                    .0
+                    .inner()
                     .shallow_clone()
                     .to_kind(tch::Kind::BFloat16);
                 let w3_tensor = self
                     .weight
                     .gate_w
-                    .0
+                    .inner()
                     .shallow_clone()
                     .to_kind(tch::Kind::BFloat16);
 
@@ -236,9 +96,7 @@ impl Expert<TchTensor> for TorchFFN {
         let module = self.load_module();
         let guard = module.lock().unwrap();
 
-        log::debug!("input shape={:?} dtype={:?}", x.0.size(), x.0.kind());
-        log::debug!("weight {}", self.weight);
-        let res = guard.forward(&x.0);
+        let res = guard.forward(&x.inner());
         TchTensor(res)
     }
 
@@ -247,8 +105,8 @@ impl Expert<TchTensor> for TorchFFN {
     }
     fn shape(&self) -> ExpertShape {
         ExpertShape {
-            dim: self.dim,
-            hidden: self.intermediate_dim,
+            hidden: self.dim,
+            intermediate: self.intermediate_dim,
         }
     }
 
@@ -281,7 +139,8 @@ mod test {
     extern crate test;
 
     use crate::{
-        ffn::{EkTensor, Expert, ExpertWeight, expert_torch::TorchFFN},
+        backend::EkTensor,
+        ffn::{Expert, ExpertWeight, expert_torch::TorchFFN},
         x::{self, test_root},
     };
