+++ ek-computation/src/lib.rs
@@ -1,13 +1,13 @@
 #![feature(pattern)]
 #![feature(test)]
 
+pub mod backend;
 pub mod controller;
 pub mod ffn;
 pub mod onnx;
 pub mod proto;
 mod schema;
 pub mod state;
-pub mod tch_safetensors;
 pub mod weight;
 pub mod worker;
 pub mod x;
